\name{RcppProgressExample-package}
\alias{RcppProgressExample-package}
\alias{RcppProgressExample}
\docType{package}
\title{
This package is an example of how to use RcppProgress in your own package.
}
\description{
This package is an example of how to use RcppProgress in your own package.
}
\details{
\tabular{ll}{
Package: \tab RcppProgressExample\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-04-17\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Karl FORNER

Maintainer: Karl Forner <karl.forner@gmail.com>
}


\keyword{ package }
\seealso{
\code{\link[RcppProgress:RcppProgress-package]{RcppProgress}}
}

